Improve usability in Endpoints Management Dashboard
Better handling creation and deletion of services accounts so that no deadlocks take place (e.g. a service no more cannot be disabled because its account was not properly created). Same with an account shared by several services, which should not be deleted if there are still services making use of it.